Moved endpoint from BudgetController to PaymentsController
