+++ Pinionszek_API/Services/DatabaseServices/BudgetService/IBudgetApiService.cs
@@ -8,8 +8,6 @@ namespace Pinionszek_API.Services.DatabaseServices.BudgetService
         public Task<Budget?> GetBudgetDataAsync(int idUser, DateTime budgetYear);
         public Task<(string?, int?)> GetFriendReceiveNameAndTagAsync(int idSharedPayment);
         public Task<(string?, int?)> GetFriendSenderNameAndTagAsync(int idSharedPayment);
-        public Task<UserSettings?> GetUserSettingsAsync(int idUser);
         public Task<IEnumerable<Budget>> GetBudgetsAsync(int idUser);
-        public Task<IEnumerable<GeneralCategory>> GetDefaultGeneralCategoriesAsync();
     }
 }
